requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
shiboken6==6.4.2
